Bump mvn extensions
When running `rmvn` from https://github.com/jruby/jruby-openssl, I get
source control changes where the `.mvn/extensions.yml` file in that
repository is overwritten with the `.mvn/extensions.yml` file in this
repo.

This is because of this code:

https://github.com/jruby/ruby-maven/blob/2cab8ac4fae8d6bc242e4893c705dde91667f806/lib/ruby_maven.rb#L65-L72

In my understanding, that essentially changes the versions of extensions
that get used by `mvn`.

So even though I'm not totally sure how all this works, I think it would
make sense to sync the files and bump the versions here too.